SDC-4220. Custom Mask Not Mask Correctly Applied.
There was a code path which was checking whether the maskingCharacter and realCharacter are not same if it is a non mask character (x)
and does not append those values to the final string. This patch removes that.

For Custom Mask, the below is the rule:
'#' character in the mask string preserves the original character in the field value string, rest of the field value characters are replaced by
the corresponding positional characters in the mask string (The total length of the resulting string is min(mask string length, field value))

Change-Id: Ib4045cc47c497cd7ddadb43136df8e8098b0e2aa
Reviewed-on: https://review.streamsets.net/4949
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>