SDC-3880: Ensure Hive Query Executor and Mapreduce Executor work with MapR
This is pretty much a bulk update required to get the new executors
working on top of mapr:

* HDFS Executor - I've disabled it as it should be subclassed into Mapr
Executor and then some of the implementation details would have to
change (the ACLs seems to work differently). We'll do this one as part
of subsequent JIRA.
* Mapreduce - I had to add the protolib to 5.1 and 5.2 stage libraries.
Otherwise it seems to work out of the box even with current mapr-setup
script (e.g. no new dependencies needed).
* Hive executor - Seems to work out of the box.

Change-Id: I0f97284ae0e395ec411e6a4859f2449982c0b1ce
Reviewed-on: https://review.streamsets.net/4992
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>