SDC-3852.  Google Bigtable Destination.
--Preliminary--

Setting up Bigtable is similar to the HBase, but you need to
add some fields related to your Google Cloud Platform setup.
InstanceID, and ProjectID need to be coordinated with what’s
specified in the BigTable Console.

The tests - BigtableTargetIT - are under development.

Google Bigtable requires credentials to be able to connect.
Credentials are created by Google Cloud Platform users by
running “gcloud auth” command.

An environment variable (GOOGLE_APPLICATION_CREDENTIALS)
is required for Streamsets.  This environment variable
must point to the credentials file and this variable
be set up in Datacollector's environment.

“Row Key” is a single field or if "Composite Row Key" is checked
the user can ente a list of fields which are concatenated and
treated as a text string and inserted into Bigtable as the
record's rowkey.

“Default Column Family Name” is optionally specified - column
family can be specified in the Column field with cf:qual
style syntax.  If every Column field has a column family specified,
and “Explicit Column Family Mapping” is checked,
“Default Column Family Name” can be blank.

If “Ignore Missing Fields” is checked, records which are missing
fields will not be treated as errors.  Otherwise, the stage's
On Record Error processing will apply.

“Create Table and Column Families” - when not checked, if
the table is not available or referencing a table that does
not exist results in an error.  Also, when this is not checked,
referencing a column family that does not exist results in
an error.

The “Default Column Family Name” interacts with the
“Explicit Column Family Mapping”.
When “Explicit Column Family Mapping” is checked, any column
family which prefaces the qualifier will be used (and created
if "Create Table and Column Families" is checked) in creatingd
the column identifier.  If"Explicit Column Family Mapping" is
not set, the "Default Column Family Name" is used regardless of the
syntax of the qualifier - the column family is replaced with the
default.

Change-Id: Iac86aeb9d3dc3718032f17db699b3d208c4e1f49
Reviewed-on: https://review.streamsets.net/5018
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>