SDC-4270: CDH: Cannot enable DPM, permission denied
The issue is that the original configuration had type "string" which
means that user is responsible to creating the directory manually. I've
changed the type to path while using type localDataDir. CM guarantees
that this directory is created and owned by the sdc user and hence we
should no longer see this error.

There is however an upgrade impact - people who have changed the
configuration before, will have to manually edit the new config. For
people who did not changed the default before, everything will work
properly. Since the amount of people using DPM is fairly small, I'm
hoping that it will be acceptable.

Change-Id: Ib822619393aaac8baf7d65b324c5a79a101b0755
Reviewed-on: https://review.streamsets.net/5037
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>