SDC-4354. Jython processor converts FileRef object to String on output.
This patch adds the conversion of ScriptFileRef to FileRef specifically to jython and javascript processor. The conversion
To fields generically happens in covertToPrimitiveObject in ScriptObjectFactory but for jython and javascript this is overriden
specifically, so this patch adds the conversion to jython and javascript processor.

Change-Id: Ib70688b7c7513df269b2e722b3a90d45aafa6b62
Reviewed-on: https://review.streamsets.net/5081
Reviewed-by: Jonathan Natkins <natty@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit 41bce7936bd25f103381bf6aed98173260b624a8)
Reviewed-on: https://review.streamsets.net/5095
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>