SDC-4474: Validate proper defaults in Avro schema
I've noticed a user who were using incorrect default values in Avro
schemas - validation issues were printed on the stdout, so outside of
reach. Avro have a very handy method setValidateDefaults that if used
will throw an exception that we can propagate as validation error.

There is however one problem - we were using this method in the past and
then stopped because it's not available in Avro version 1.7.3 that we
have to keep compatible with because of MapR. I've overcome the issue by
putting a try-catch block that will simply ignore NoSuchMethodError
exception when older avro library is found on the path.

Change-Id: I640f95ccae459706e60413be8a793ec0babcea30
Reviewed-on: https://review.streamsets.net/5236
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>