SDC-4446. HDFS destination does not seem to be flushing at the end of every batch.
This patch fixes the case where hflush is not called for each batch completion. We added a wrapper output stream
for FsDataOutputStream which when we call flush will call hflush as well.

I manually verified this, after this patch i am able to see the data on the _tmp file after batch completion.

Change-Id: I72ad6dc08ebfd8e3a1297c3cc6e017700eec9b67
Reviewed-on: https://review.streamsets.net/5244
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>