SDC-4409. Hadoop FS destination may not be renaming all TMP files on recovery
rename all temp files while initializing the pipeline if dir path is given in the config

this issue arise, if we have the following tmp files after the pipline crashes:
/tmp/out/20161011/us/94086/_tmp_sdc_123456.txt
/tmp/out/20161011/korea/135789/_tmp_sdc_123457.txt
the pipeline restarts and the renaming tmp files on recovery in /tmp/out/20161011/us/94086/ will never happen if we have no records writing to that directory.

so, scan all subdirectories and rename all tmp files on recovery while starting up the pipeline.

if "directory template" is in the record header, this will be ignored.

Change-Id: Icceb085caaab956d50244fd81662ca2f1a8a5814
Reviewed-on: https://review.streamsets.net/5274
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>