SDC-4437 LDAP Authentication should support members being defined by uid
Using our current implementation, I added a new config "roleFilter".
How to use the filter is same as ldaptive jaas
http://www.ldaptive.org/docs/guide/jaas.html

If the filter has {dn}, we replace {dn} with user's full DN. If {user}, apply uid.
If roleSearchFilter is not set in ldap-login.conf, or roleSearchFilter is set but
doesn't contain {dn} or {user}, we apply the original behavior(search by full DN).

Added more integration test to test this behavior.

Change-Id: I9894a5a766c7bfacdcfab61ee22d880ac932d8a0
Reviewed-on: https://review.streamsets.net/5277
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>