SDC-4510. Cannot start a pipeline with dir origin the second time.
isFileFromSpoolerEligible(File spoolerFile, String offsetFile, String offsetInFile). This function decides whether the spooler listed files are eligible for produce function to parse for more records.

In this function we are making a bunch of comparisions with the fileName, but we recently changed the offsetFile parameter to be the full file path, so these checks need to be updated to compare the full path, rather than fileNames.

Because of one of the comparison we repicked a file again which has already been processed fully and offset set to -1 and parsers were failing to read from offset -1.

This patch fixes the isFileEligibleForSpooler to use the whole file path for comparison rather than just the file name.

Change-Id: I72dfcab14ec0d8c42fb7cbd07e6316127023df5c
Reviewed-on: https://review.streamsets.net/5272
Reviewed-by: Bob Plotts <bob@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>