SDC-4513 str:truncate
StringEL.truncate("", endIndex) causes StringIndexOutOfBoundsException
because we use substring(0, endInex) to implement truncate, and
we are not checking if string is longer than endIndex.
Added fix to return the original string if endIndex is bigger than
the length of the string.

Pipeline always fails sdc:truncate while initializing and validating the
configurations. While validating if EL is valid, we use empty record
so that record:value() returns null, and the null gets converted
to empty string "". Then it hits the above bug.

Change-Id: Idf7bc8bb285827ab79dda9033f20858a59b12b84
Reviewed-on: https://review.streamsets.net/5315
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>