SDC-4484. Provide a way to specify extra conditions on offset columns.
This patch adds an extra configuration named "Extra offset column conditions" through which user can restrict the query to
specific extra conditions other than the last offset (i.e the conditions gets specified in the where clause as and AND condition
along with the conditions for last offset already applied)

The most common use being if a table has a transaction time column (which is not unique) and records the current timestamp (till milliseconds)
as the transaction time. There might be multiple active transactions which are getting inserted (at the same millisecond) when ingestion happens
(and we should try to read all the rows with same transaction time), it may so happen that not all records are inserted for the current active
timestamp and we might skip to the next offset (based on > last offset) in the next batch and skip some rows in the current timestamp.

So avoid that user could basically specify transaction time < ${time:now} to not read the records in the active timestamp
and strictly read less than it to allow for all the transactions in the current time stamp to be recorded.

Also made a fix to make sure to store date/time/datetime in string based on a format so the query can do > lastOffset gracefully

Change-Id: Ic708d9d959cbdd8e4d78851012b36d1df442b7e6
Reviewed-on: https://review.streamsets.net/5553
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>