SDC-4628.  Bigtable target record counter may keep increasing.
Changed two things about processing records which have a
field that has a conversion error or have fields which
are null.

Any record with a conversion error will be sent to the
On Record Error destination regardless of the setting
of the “Ignore Missing Fields” switch.

Records containing a field that has been nulled is
now handled as follows:

When “Ignore Missing Fields” is not checked in the UI, any
record which has a field that is null is sent to the
On Record Error destination, and no fields will be added
to Bigtable.

When “Ignore Missing Fields” is checked (enabled),  any
record in which ALL fields are null is sent to the
On Record Error destination.  If at least one field from
the record will be sent to Bigtable, the other fields
which are null will be silently ignored (nothing is sent
to the On Record Error destination).

Also enhanced some error messages as suggested in SDC-4602.

Change-Id: I336ea1745d204a44aebba5c9b3534ab989c26f38
Reviewed-on: https://review.streamsets.net/5560
Reviewed-by: Virag Kothari <virag@streamsets.com>