SDC-3179.  MapR-DB JSON table destination
This destination handles inserting documents
into a MapR DB JSON document database,
version 5.1 and 5.2 are supported.

Currently this destination takes all the columns in the
incoming record and forms them into a JSON document, and
Inserts the document into the database.

Date, Time and Datetime fields are represented as long
values in the JSON document.  If a Date, Time or Datetime
field is used as the _id column, it is also stored as a long
value.  If you require a different representation for Date
fields, use a fieldTypeConverter to create the appropriate
String representation of the date.

The default character encoding for the JSON document
is UTF-8.

There is a UI feature to change that behavior to use
InsertOrReplace, which will simply replace the document
with the corresponding key, if there is one in
the database already.

The UI features are:

Table Name - this is the MapR DB table name.
Create Table - if the table does not exist, MapR DB JSON
will create it.
Field to use for _id - specify the key field - this
field acts as a primary key and should be a
unique data value.

Use Insert or use InsertOrReplace - this listbox lets
the user select:
InsertOrReplace (new documents which have duplicate _id
fields replace the previous versions).  Any failure calling
this routine will result in a StageException.
or
Insert (when there is a duplicate _id field, the record
is sent to On Record Error destination).

Change-Id: I479e92591832b4bd1d52c7f997d47720d9f57c24
Reviewed-on: https://review.streamsets.net/5563
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>