SDC-4769. Whole File Format: Provide a way for rate limiting the stream read.
This patch adds a rate limiting wrapper stream which can limit the rate with which a stream can be read.
Also added a configuration for whole file format origins to specify a rate limit which allows DataUnitsEL.
Made changes to all whole file format supported origins to consume the rate limit.

Change-Id: I2b3b3281c21a5853980149a995b928ed51c58e88
Reviewed-on: https://review.streamsets.net/5790
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>