SDC-4789: Add support for push origin to ProductionPipelineRunner
This the "break the world" kind of a patch. I've refactored the normal
execution path to be special case of a multi-threaded path. I've
intentionally left a lot of things that needs to go in to support
multi-threaded pipelines out of this patch to keep it simpler-ish for
review.

This patch should be no-change for existing pull origin and should
provide basic runtime support for push origin (tested on quick&dirty
refactoring of data generator origin).

Change-Id: Ia30f9490cf3069ea5b03af697425462654d50f1d
Reviewed-on: https://review.streamsets.net/5782
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>