SDC-4785: Create MetricsConfigurator methods for reusing metric objects in multithreaded pipelines
This patch refactores the MetricsConfigurator that is used by pretty
much everyone to register metrics objects to allow metric-reuse. The use
case is to have one single counter, histogram, ... for all instances of
given stage so that we can easily report unified metrics in multi
threaded pipeline.

This is a first step for introducing metrics to the multithreaded
pipelines. The next will be to introduce way to also report metrics
separately for each individual pipeline runner. However that is a future
work and hence it's not present in this patch.

Change-Id: If8d8e89b3bf567c47054d9c347d45e6a77124372
Reviewed-on: https://review.streamsets.net/5860
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>