SDC-4870.  Enhance MapR JSON Destination to handle BINARY row keys.
The MapRDB JSON Document database supports both String and BINARY
keys - this ticket is to enhance Data Collector’s MapR DB JSON
Destination to support BINARY keys.

The Row Key type - BINARY or String - is determined by the data type
of the field chosen to be the Row Key.  If the type is byte array,
the Row Key is added to the document as a BINARY key.  If the
row key data type is String, the field is used by MapR DB JSON
as a string key.

No data type conversions are performed by the
MapR DB JSON Destination.  When necessary, use a
FieldTypeConverter to convert the datatype of the
key field to byte array (or String).

When using the BINARY key feature, changing the key’s
data type or size after records have been added to MapR DB
will not result in an error, but probably will not
give the desired results.

It's possible that you will want to convert DATE, TIME or DATETIME
fields to an epoch in milliseconds or a different format
date string - use a FieldTypeConverter for this conversion as well.

Change-Id: Idb7d3da57430f8fcc3e9808269f9773789f9a268
Reviewed-on: https://review.streamsets.net/5918
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>