SDC-4868. MongoDB Oplog Origin.
Use Oplog in MongoDb to produce records, we use the tailable cursor to tail oplog and get records using mongo collection find api.
Users can filter records based on what operation types they are interested in. We use the oplog ts field as the offset field.
Also we populate the "ns" and "op" field in the header, so it is consistent with other origins like JDBC.

The op type in the header will be one of the following "i", "d", "u", "c", "db", "n".
Inorder for users to do the required option on the target, users should use something like static lookup to convert these optypes
to correponding verbes for the target needs (For ex: Mongo target requires SDC.MONGODB.OPERATION)

Change-Id: I160fef9ac73b38cf93277fd0ab48abb45cb77fec
Reviewed-on: https://review.streamsets.net/5940
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Ji Sun Kim <jisun@streamsets.com>