SDC-4885: PushSource.getNumberOfThreads() is called before init()
The issue why we were calling the PushSource.getNumberOfThreads() before
init() is that the way code is currently written is that the whole
pipeline including all runners is created before any initialization code
is running.

I've significantly refactored the Pipeline class to change this behavior
and:

1) At the time of creating the Pipeline, we create only one runner. This
means that we can easily verify that all components are creatable.
2) On init() after the origin is initialized, we dynamically create the
remaining runners.

Change-Id: Ic8b317380c17e3b059f4e0191125c955114b584c