SDC-4884: Stopping pipeline with push source can lead to IllegalMonitorStateException
This one was easier to solve as the underlying class DataStore is
already doing synchronization to protect against multiple-access. Sadly
one of my earlier refactoring broke it's use in OffsetFileUtil by not
properly releasing it upon read. This patch add the release back which
seems to fix the issue with stopping pipeline.

Change-Id: If6d4f58e5756aec5ca1aa8b6b9b6835a11aea152
Reviewed-on: https://review.streamsets.net/5936
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>