SDC-4825.  time:extractStringFromDate must support time zone conversion.
Created a new method: extractStringFromDateTZ.  This method accepts a
time zone and processes the provided date object in accordance with the
specified time zone.

This method takes 3 parameters - first is a Date object.

The second parameter is the text description of the time zone eg.
‘America/New York’ or ‘Asia/Riyadh’ for example.

For users working with US dates and times, the EST5EDT time zone format
is supported, and daylight savings time adjustment works.

Numeric-only time zones such as ‘-0500’ or ‘-08:00’ are not permitted
and will return an empty string.  Numeric time zones need to be prefaced
with ‘GMT’ for example: ‘GMT-0500’ or ‘GMT-08:00’.  For many regions of
the world, it is not recommended to express time zones this way as they
do not correctly handle daylight savings time.

It is possible to use Short Time Zone IDs such as EST, CST.  They are
deprecated - and will have issues since some of the 3-letter IDs
represent multiple time zones.  CST, for example, is valid for US
Central Standard Time and China Standard Time.   Per this link,
3-character time zone IDs are deprecated and should be avoided:
http://docs.oracle.com/javase/7/docs/api/java/util/TimeZone.html
see the section titled: “Three-letter time zone IDs”

The third parameter to this method is an output date mask to describe
the format of the output string.  Use this link
http://docs.oracle.com/javase/7/docs/api/java/text/SimpleDateFormat.html
to see the fields which are available for the date mask.

Example EL expression:
${time:extractStringFromDateTZ(record:value("/dateTime"),
		"America/New_York",  "yyyy-MM-dd HH:mm:ss zzzzz")

The Date object representing 2016-08-28 22:25:13 will be converted to
the string “2016-08-28 17:25:13 Eastern Daylight Time"

In this example, Eastern Daylight Time is added to the output
string as the time zone; by using the time zone ID
(America/New_York), the conversion is able to distinguish whether the
geographic region is in daylight saving time and will adjust the time
and the time zone name accordingly.  The time zone is not required to be
in the date mask.

Overview of DST by country:
https://en.wikipedia.org/wiki/Daylight_saving_time_by_country

List of valid time zone IDs:
https://www.vmware.com/support/developer/vc-sdk/visdk400pubs/ReferenceGuide/timezone.html

Change-Id: I73bbd980c00d6e5cf4a14065853e8bc66f7d1b1b
Reviewed-on: https://review.streamsets.net/6031
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>