SDC-4986.  enhance binary key logic.
the following types are supported as binary keys:
Short, Integer, Long.
Date, Datetime, and Time are converted to epoch in MS prior
to being converted to binary.

String is converted to a byte array.
Byte array is passed through and is used verbatim.

if the row key field is a byte array in the incoming record,
it is used as a binary key, regardless of the UI setting.

Change-Id: I74a293b11a9904c1f43834de5172fb2a1e6fb987
Reviewed-on: https://review.streamsets.net/6047
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>