SDC-4994: TestStandaloneRunner is frequently failing
The reason behind the test failures was that the DataStore class was not
synchronizing properly. We have a static map that is suppose to hold
lock for given file path (offset file), but it could easily happen
that there were multiple locks for the same file which lead to various
random exceptions.

This patches fixes several issues that are all highly connected, so I'm
not submitting them independently:

* This patch reverts my previous patch SDC-4884 since that fix was not
  correct and the fact that it "helped" is more coincidental.
* Added abstraction on top of the lock to count how many times it's
  "used" - not necessary locked, but how many threads holds the instance
  of the lock for various actions. And we're removing the counter-lock
  only in case that 0 threads hold the instance to prevent creation of
  two locks for the same file.
* The operation exists() on the DataStore was previously not
  synchronized, however it have to be because it is not ready only - it
  will actually try to "recover" failed write and I've seen scenario
  when it tried recover write that was just in the middle of commit().

I've run the TestStandaloneRunner 100 times to make sure that the
flakiness is gone now.

Change-Id: I89dba90f565b1934d85095c8407332b81beca1d6
Reviewed-on: https://review.streamsets.net/6064
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Reviewed-by: Hari Shreedharan <hshreedharan@streamsets.com>