SDC-5003: Synchronize access to pipeline singletons
While doing some testing of multi-threaded pipelines, I've realized that
we can run both the error stage and stats aggregation stage in parallel
which is not in the contract with those stages. Hence we should
synchronize access to them for now and consider instantiating more of
them and have one for each pipeline runner.

Change-Id: I1572a57376c9f5cb6b01581b7b8070ee001ffaf9
Reviewed-on: https://review.streamsets.net/6074
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>