SDC-5005: Using STOP_PIPELINE OnErrorRecord policy doesn't work for multi-threaded pipelines
Our current code for single-threaded pipelines is written in a way that
it's assuming that any exception from processing should shut down the
pipeline. This is a bit harder to accomplish in multi-threaded pipeline
since the origin is controlling the execution.

This patch changes the behavior of processBatch() handler to cache the
exception that was thrown and signal origin that it should stop
processing request. This way all other batches that might succeed are
properly finished. Then after PushSource.produce() is done executing we
re-throw the exception.

In case that multiple threads/runners throws exception, only the first
one will get propagated to the UI. The rest will be available in the
logs.

Change-Id: I7a034b90c6bde3ed37ef0d9606f6b7a7158abc53
Reviewed-on: https://review.streamsets.net/6076
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>