SDC-5017. Cache Result Set to avoid frequent table querying.
Introduce a cache to store results set especially for when we switch tables so as not to query the database frequently. But we will end up having multiple open statments/result set, in order to control how many open statements can be there we use expose a result set cache configuration which user can manipulate.

Change-Id: I643ad9bcf0faeb80cf2bda5ba30ca9bd40f9f19b
Reviewed-on: https://review.streamsets.net/6083
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>