SDC-5021: Incorporate API-82 to Datacollector code base
I needed to do few clean ups:

1) EventCreator was assuming that Context.toEvent will always work which
is not valid assumption for PushSource. Refactored the class to take
both Stage.Context and then ToEventContext. I've kept original
constructor to avoid changing all the places where it's used already.

2) DefaultErrorRecordHandler made similar assumption as EventCreator
just for Error. Again the assumption is invalid for PushSource, so a new
constructor was provided. Also kept original constructor so that we
don't have to change all 57 classes that use it.

3) HDFS FSFormatHelp was at few places using Stage.Context instead of
Target.Context that.

Change-Id: I6b5ac8ab50d4788d0342de63d11f0fa0b5b59ce9
Reviewed-on: https://review.streamsets.net/6090
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>