SDC-5016: Compile dependency on kafka-clients in sdc-kafka-10 module breaks cluster mode on hdp 2.5
Configured the kafka-clients dependency as runtime (same as kafka
itself) and explicitly excluded it in the HDP profile (similarly as
we're doing for main kafka jars).

I've validated the resulting build of stage lib that it does not contain
the different version of kafka-clients. I did not run a cluster mode to
validate further though.

Change-Id: I0032807f5270e2f9c61fef69ecfd8129613786b8
Reviewed-on: https://review.streamsets.net/6091
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Virag Kothari <virag@streamsets.com>