SDC-4851: SDC should produce a heap dump on OOME
Added new property that defaults to true - if set it will auto-add
arguments for generating heap dump on OOM. By default the heap dump will
be stored in SDC_LOG/heapdump directory - I've picked log dir as I'm
assuming that people will be actively looking to this directory for
space consumption. User can of course override the location in the
configuration file.

Change-Id: I7e2ff1ecb1e25ad3616ded16ff04f89d79155e78
Reviewed-on: https://review.streamsets.net/6093
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Alejandro Abdelnur <tucu@streamsets.com>