SDC-4781.  Add MapR DB JSON table Origin
This Origin reads a MapR DB JSON table, using “_id” as the index, and
parses the documents into Records then provides batches of Records
to the next stage in the pipeline.

This Origin supports both String and binary keys.  The key types cannot
be intermingled within one table, that is, you cannot have some JSON
documents with the String value for the key field you select, and some
documents with a binary key value.  Inserting data works correctly,
retrieving a document which has a String key as binary or retrieving
binary as String will result in a StageException.

Binary keys should also have the same width for all the rows in the
table.  Data for binary keys should be of a numeric data type -
String will throw StageException.

For binary keys the data element is converted to binary as the
appropriate width data item, (8 bytes for Long4 for Integer, 2 for
short), and added to the document as the _id field.

String keys use the String value of the data items, and sort
alphabetically, from left to right. If Date, Time, and DateTime
are used, they are converted to milliseconds from the epoch
then are used as Long values.

When data is added to the JSON table by Data Collector MaPR JSON
Destination or another application, the data can be inserted as
binary, and will preserve the sorting sequence of a numeric
column such as Long or Integer.  Date, Time and Datetime can be
supported by converting them to epoch time in milliseconds.

MapR DB JSON Origin dynamically figures out if the _id key column is
String or binary.

When passing data to the next stage in a  pipeline, if the key was
binary, the value for the _id field is replaced with the string
representation.  If the _id column is unwanted remove it with
a subsequent FieldRemoverComponent.

UI components:
Table name - name of the MapRDB JSON table to read.

Starting Key Value -- this value indicates the starting value for
document selection. documents with a key value equal to and greater than
this value are selected from MapR DB.  This value is converted to
the appropriate size (byte width) for binary keys.

Change-Id: I11bb0aabeb44e45ffc2fbed6aca92bd53e600299
Reviewed-on: https://review.streamsets.net/6104
Reviewed-by: Virag Kothari <virag@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
(cherry picked from commit ce3d83ed67a788a393b0b73e9e7c9827a75ba7d8)
Reviewed-on: https://review.streamsets.net/6157
Reviewed-by: Bob Plotts <bob@streamsets.com>