SDC-5013. Enable Invokedynamic support in Groovy Processor
Enables use of the indy (invokedynamic) compiled Groovy libraries.
Because JSR-223 does not provide a way to configure script engines
we have to create our own ScriptEngineFactory implementation. We use
a new script engine short name "groovy-sdc" to load it in order to
avoid loading the standard (default) GroovyScriptEngineFactory.

This is because we must enable the indy compiler flag otherwise
only core groovy libraries will use invokedynamic and not user defined
scripts.

Change-Id: I9329eb7aeee2d36d863030d36ffbb483a56e1da1
Reviewed-on: https://review.streamsets.net/6139
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>