SDC-5058. SFTP origin throws exception when processing zipped files.
This patch fixes the case where a compressed archive/archive files are used with FTP source where previously we assumed
offsets are a Long, for archive files it is actually a json with inner file of the archive along with offset.

There is a change in behacior too where we have always been storing the offset of a file and move to next file
and we never stored once the file is complete, from this patch we also store -1 in the offset if the file is
completely parsed. (It could happen that we read the last offset and we never try to read beyond that, in
that case expect an empty batch).

Change-Id: I77c04df9e55deb8ccb53761b3aa43423fb5d6bb1
Reviewed-on: https://review.streamsets.net/6146
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>