SDC-5060: Introduce simple script to validate versions of pulled dependencies
We had few issues in the past when we've mistakenly shipped one
dependency (netty) with multiple versions inside one stage library which
broke that stage library. The idea here is to introduce a script that we
could run on a release to validate that we're not shipping any duplicate
dependencies.

The script already found several stage libraries that ships duplicate
dependencies that will be fixed in separate JIRAs.

Change-Id: I093913d51e099debdfa0db56c3dfd81ba6d3cae5
Reviewed-on: https://review.streamsets.net/6147
Reviewed-by: Dima Spivak <dima@streamsets.com>