SDC-5111: Rename pipeline: Pipeline EL doesn't seem working
Pipeline EL did not worked for PushSource because there is a special
initialization code that stores information in ThreadLocal variable.
This code have run only inside Runners for the "main" Thread.

I've added the initialization routine also to the
ProductionPipelineRunner so that we created proper content of this
ThreadLocal variable even for PushSource.

There is one gotcha though - the PipelineEL will be only available
within batch context - e.g. after method startBatch an before
processBatch.

Change-Id: Ie6c5261dfe266eea7022f50e99b8178ea3879db2
Reviewed-on: https://review.streamsets.net/6303
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>