SDC-5170. TestStandaloneRunner is flaky with 'runner closed' issues
Introduce a config runner.expiry.initial.delay for the initial delay before
executing the runner expiry thread. The default is 30 mins as there is not much
advantage on running this thread as soon as sdc starts. Also there is hope to
fix the flaky TestStandaloneRunner where the runners would sometime close very soon.

Change-Id: I741e03d40012a97dfeda0913b3d00f2a3b9946c9
Reviewed-on: https://review.streamsets.net/6319
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>