SDC-5196. Hive Target takes lot of time for tables with more number of partitions.
This patch basically removes the eager loading of partition locations and move it to be more lazy loading. We first simply lists the partitions and then were doing 'DESC FORMATTED' query on partition to basically extract the location information of the parition. The location information mainly used for validation purpose. This patch changes this to be lazy loading to extract partition location, thereby we simply extract partition location when we need it.

Change-Id: Ie26a3520fca0d5713865452f003ab2afb23e97a4
Reviewed-on: https://review.streamsets.net/6372
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
(cherry picked from commit 935d1611c9b878da28f780e7d33cb42f14fffda6)
Reviewed-on: https://review.streamsets.net/6380
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>