SDC-5195: Null field output by Field Renamer
Core of the issue is that our method Record.getEscapedPaths() will
auto-double three special characters - /, [ and ]. But then the method
Record.get(Path) will not de-double them. So person who is iterating of
the getEscapedPaths() can in fact get a path that doesn't exists.

I was trying to find a reason why we're doubling the three characters
and as far as I can tell, this seems to be an old concept that we had in
the code for "escaping" that was eventually superseded by the idea to
enclose field names in quotes.

Change-Id: Ia65a6d50213a0c9ab7f3bb3349ab66ff1d2babed
Reviewed-on: https://review.streamsets.net/6371
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
Reviewed-by: Harikiran Nayak <hari@streamsets.com>