SDC-5328. Record.getEscapedFieldPaths() does not return unescaped name when \ or ` is present.
This patch fixes the case where a \ or ' is present in the field path and is escaped a special character when record.getEscapedFieldPaths is called and when the fieldPath is used for lookup using record.has it does not find the path in the map. The solution is to simply unescape the above mentioned characters so map lookup and basically lookup the values in its keys.

Change-Id: If0741699154989ba47a1691b7e5ea395aa0b2eea
Reviewed-on: https://review.streamsets.net/7374
Reviewed-by: Jarcec Cecho <jarcec@streamsets.com>
Reviewed-by: Madhukar Devaraju <madhu@streamsets.com>