SDC-5076: Expose configuration of spool size in S3 origin
Exposed the size of the pool to the user to let them configure how many
files they want to keep per iteration. I would appreciate feedback on
the naming of the property as I'm not that good in public names :(

Also I'm not entirely sure how to properly test this change, so feedback
is welcome.

Change-Id: I036e718f64e518ebd78528ae7dec8174d752f106
Reviewed-on: https://review.streamsets.net/6446
Reviewed-by: Santhosh Kumar <santhosh@streamsets.com>
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>